Update: improve normal, warning, and error messages
Many of the warnings should actually be reported as errors.
Add error and warning prefixes to the log file to make searching a large operation set for errors or warnings easier.
Use custom message functions to simplify (and make more consistent) the output and logging functionality.

Fix minor warning or error reporting mistakes.