Update: move "find" to a variable and add "sort" command as optional
When comparing the results of a dedup process to another identical one, for testing, I noticed that the find command has different find orders.
Pipe the find command to a sort command to ensure consistent results but do so only if "sort" command is available.

Move the find command to a variable to make it easier to change if some user wanted to.