Feature: add saf rename directory script
This is the original, unmodified, script that I used to rename directories generated by SAFCreator into names that matched a distinct field in our data set, called "Serial ID".
I intend to update this script and formalize it based on the practices I am now following in this repository.
This has been uploaded to preserve its original state prior to any of the planned cleanup/formalization changes.