Update: improve output and logging using padding, treat invalid directory names as warnings, and handle error when directory name is not a number
Use padding in the output and logs to improve readability (and remove extra new lines).

Invalid directories should be a warning and not an error, in the event that the source directory contains other (irrelevant) directories.

When the directory name is not a number, prevent bash from throwing an error.
Using `let x=$y` requires `$y` to be a number.
Pre-check `$y` to be a number and thrown a more controlled error message when `$y` is not a number.