Update: add handling of 'invalid' file for problematic PDFs
Implemented using a single function that can be reused between 'duplicates' and 'invalid'.
Includes preliminary support for configuring whether or not to enable/disable a processor: 'duplicates', 'invalid', or 'missing'.
For now, these can be manually disable by setting the process value to 0 in the variable initialization area of `main()`.